requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
python-dotenv>=1.0.0